refactor: extract _UIA_SCRIPT to _uia_script.py to pass file-size gate
_windows_uia_backend.py exceeded the 1000-line Python limit (1032 lines).
Extract the ~293-line PowerShell UIA script constant into _uia_script.py,
bringing the main backend file down to 744 lines.

Use importlib.util.spec_from_file_location for robust loading that works
in both normal imports and dynamic exec_module contexts (e.g., tests).

Refs: core#1894
